Separate production and development mode

+++ src/index.ts
@@ -3,8 +3,13 @@ import * as api from './api'
 
 async function init() {
   try {
-    console.info("Connecting to database")
-    await db.connection.sync({force: true})
+    if (process.env.NODE_ENV == "development") {
+      console.info("Connecting to database in DEVELOPMENT mode")
+      await db.connection.sync({ force: true })
+    } else {
+      console.info("Connecting to database in RELEASE mode")
+      await db.connection.sync()
+    }
     console.info("Starting API endpoint")
     await api.default()
   } catch (e) {
